requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0
